requirements.txt
discord.py
prometheus_client
